backend/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.1
sendgrid==6.11.0
pyjwt==2.8.0
mysql-connector-python==8.3.0
python-dateutil==2.8.2
requests==2.31.0
werkzeug==3.0.1 